+++ RUFAS/biophysical/manure/manure_manager.py
@@ -59,7 +59,7 @@ class ManureManager:
         A list defining the execution order of processors.
     """
 
-    def __init__(self) -> None:
+    def __init__(self, intercept_mean_temp: float, phase_shift: float, amplitude: float) -> None:
         self._om = OutputManager()
         self._manure_nutrient_manager = ManureNutrientManager()
 
@@ -77,7 +77,9 @@ class ManureManager:
         processor_connections_by_name = self._validate_and_parse_processor_connections(
             processor_connections_input, processor_configs_by_name
         )
-        self._create_all_processors(processor_connections_by_name, processor_configs_by_name)
+        self._create_all_processors(
+            processor_connections_by_name, processor_configs_by_name, intercept_mean_temp, phase_shift, amplitude
+        )
         self._populate_adjacency_matrix(processor_connections_by_name)
 
         self._validate_adjacency_matrix()
@@ -672,6 +674,9 @@ class ManureManager:
         self,
         processor_connections_by_name: dict[str, dict[str, list[dict[str, Any]]]],
         processor_configs_by_name: dict[str, dict[str, Any]],
+        intercept_mean_temp: float,
+        phase_shift: float,
+        amplitude: float,
     ) -> None:
         """
         Creates and initializes all processors based on their definitions.
@@ -692,6 +697,10 @@ class ManureManager:
             if not (issubclass(processor_initializer, Handler) or issubclass(processor_initializer, Separator)):
                 del processor_config["processor_type"]
             processor = processor_initializer(**processor_config)
+            if isinstance(processor, AnaerobicLagoon) or isinstance(processor, SlurryStorageOutdoor):
+                processor.intercept_mean_temp = intercept_mean_temp
+                processor.phase_shift = phase_shift
+                processor.amplitude = amplitude
             self.all_processors[processor_name] = processor
 
             if isinstance(processor, Separator):
@@ -896,7 +905,7 @@ class ManureManager:
             "non_degradable_volatile_solids",
             "degradable_volatile_solids",
             "total_solids",
-            "bedding_non_degradable_volatile_solids"
+            "bedding_non_degradable_volatile_solids",
         ]
 
         for name, processor in self.all_processors.items():